requirements/base.txt
fastapi==0.100.1
httpx==0.28.1
python-multipart==0.0.9
pypdf==4.2.0
tortoise-orm[asyncpg]==0.19.3
aerich==0.7.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
redis==5.0.1
boto3==1.28.62
aioredis==2.0.1